Optional check if user is member of a specific group
This adds the following command line flag `--ldap-group` to set the
group a user is checked against.
Make sure to not specify a organizational unit `ou` in your base DN
because otherwise either user or group searches will not work, at least
if `ou` is set to `people` or `groups`.